+++ src/frontend/formulation/mod.rs
@@ -12,5 +12,5 @@ pub use parser::{
     parse_expression_alias, parse_expression_binding, parse_form_or_declaration,
     parse_hard_cast_statement, parse_is_via_statement, parse_label_header,
     parse_ordinary_declaration_statement, parse_refined_declaration_statement,
-    parse_resource_header, parse_spec_operator_alias, parse_writing_alias,
+    parse_resource_header, parse_spec_operator_alias, parse_topic_header, parse_writing_alias,
 };